I noticed that the Makefile was always rebuilding the project (BAD) so I took a look to find what was wrong. I made some changes that seem to have fixed it (in Linux, at least). Namely, the linking target should be called `$(BIN)' instead of `$(NAME)' because the actual file built doesn't necessarily correspond to the NAME, but to the BIN. Also, I realized that header files should not be in the .PHONY target (:#). That was my mistake. The .PHONY target is actually used for targets that do not correspond to files on the file system (i.e., clean and default). The header files do, and when make doesn't find a rule for them it should check the file system, so they are automatically happy.
I also moved the readme.txt file to README.txt because it is supposed to stand
out in a directory listing. I was tempted to go with README.txt, but I decided
to be understanding of [n00b] Windows users that like double-clicking icons. :D

TODO: My next goal will be to get a mouse cursor drawn in game (preferrably the
native one). I don't plan to use the mouse at all in game, although it would be
nice for the GUI to be mouse aware; however, I find the way the mouse is hidden
over our game window is distracting for the user when they're interacting with
the rest of their desktop. The mouse probably isn't required in fullscreen mode
so I'll probably try to arrange it so the pointer is only visible in windowed
mode.



git-svn-id: https://intern-quest.googlecode.com/svn/trunk@53 988baf0e-e108-11dd-bd7b-b12b6590754f