Lots of changes, moving most of the global routines into the app class... >_> There is also a timer class now that is intended to count game time. I'm not sure yet how to encapsulate the frame timer, but it should be done... For now, it has begun to be moved into the app class as well.
Unfortunately, I have brutally broken this and the application dies a quick
death. No output and gdb only displays a few threaded addresses, etc... >_>
Seems over my head so I think I'm going to try to old trusted printf() or
std:cout approach to figuring out where it dies in my code.

I'm very probably doing something wrong with timers/callbacks/methods that the
compiler can't catch... >_> Yet to be determined.

At least now the code is in the repository to never be forgotten and to perhaps
revert to if i undo too much...

Note that IIRC the BASE revision for this changeset was r55 (the previous one).
So when merging, I think it's safe to assume it was based on r55, and therefore
the changeset should be just r56.



git-svn-id: https://intern-quest.googlecode.com/svn/branches/bamccaig-unstable@56 988baf0e-e108-11dd-bd7b-b12b6590754f