Merged /branches/bamccaig-unstable with /trunk with changeset 71.
Moved the draw and logic routines into the iq::app class. Beginning to think
of how to control the different states of the game. I'm thinking a QnD
solution would be to just store an enumeration state (substate, according to
the technical documentation) and have different draw/logic routines (methods)
for each. Then the main draw/logic routines (now methods) will act as switches
to pick the correct routines to execute.

An alternative that I just thought of, which should save the minor expense of
the redundant switch statements, is to execute a method to change the state,
which would not only set the game's state, but would also assign the correct
draw and logic methods to the actual draw and logic methods, which would be
function pointers... Not sure which would be more efficient/most clear.

Hmm...



git-svn-id: https://intern-quest.googlecode.com/svn/trunk@72 988baf0e-e108-11dd-bd7b-b12b6590754f