Moving my Makefile to a non-default name. The other Makefile is more automated and less likely to be broken by modifications. At the same time, it seems to have imperfections, such as not rebuilding the code when header files are modified, etc.
This way, a file needs to be specified and users of incompatible platforms are
less likely to be hounded with Makefile problems. In short, if you want a
Makefile that doesn't need to be managed you should use Makefile.old. It may
require cleaning the project directory with `make clean' after certain changes
to make sure things are properly rebuilt though.

Makefile.bam should continue to work in Linux because that's where I'm
developing it. So you might consider using it if you're compiling from Linux.

Since I'm in Linux, I plan to symlink my Makefile (Makefile.bam) to Makefile.


git-svn-id: https://intern-quest.googlecode.com/svn/trunk@81 988baf0e-e108-11dd-bd7b-b12b6590754f