Adding concepts of time to the animation class and therefore hopefully the ground work for animation. The actual API and technique is still somewhat theoretical though.
ATM, the player's walking-right animation is loaded and drawn in the center of
the screen. The actual animation part isn't working though as it seems to
always be the same image drawn.

The error/exception handling has also been reworked a little bit. As is now,
exceptions are always stack objects derived from std::exception and always
caught by reference. The majority of main() has been wrapped in a try...catch
so any exceptions thrown beneath it should be caught and reported to the user.
As for gracefully catching non-fatal exceptions and handling them, that will
have to be done at a lower level or inside of the outer try...catch, as
required.



git-svn-id: https://intern-quest.googlecode.com/svn/trunk@92 988baf0e-e108-11dd-bd7b-b12b6590754f